finance/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
beautifulsoup4
streamlit
pandas
matplotlib
plotly
seaborn